modify test to createTable
